pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yaml_config_manager"
dependencies = [
  'PyYAML ~= 5.4'
]

version = "1.0.0"
authors = [
  {name="Guyshe"},
]
description = "Small package for working with config files"
readme = "README.md"
requires-python = ">=3"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/guyshe/yaml-config-manager"
